organize code better
