+++ net-utils/src/lib.rs
@@ -4,32 +4,28 @@
 #[cfg(feature = "dev-context-only-utils")]
 use tokio::net::UdpSocket as TokioUdpSocket;
 use {
-    anyhow::{anyhow, bail},
-    bytes::{BufMut, BytesMut},
     log::*,
     rand::{thread_rng, Rng},
     socket2::{Domain, SockAddr, Socket, Type},
     std::{
-        collections::{BTreeMap, HashSet},
         io::{self},
-        net::{IpAddr, Ipv4Addr, SocketAddr, TcpListener, TcpStream, ToSocketAddrs, UdpSocket},
-        sync::{Arc, RwLock},
-        time::{Duration, Instant},
-    },
-    tokio::{
-        io::{AsyncReadExt, AsyncWriteExt},
-        net::TcpSocket,
-        sync::oneshot,
+        net::{IpAddr, Ipv4Addr, SocketAddr, TcpListener, ToSocketAddrs, UdpSocket},
     },
     url::Url,
 };
 
+mod ip_echo_client;
 mod ip_echo_server;
 pub use ip_echo_server::{
     ip_echo_server, IpEchoServer, DEFAULT_IP_ECHO_SERVER_THREADS, MAX_PORT_COUNT_PER_MESSAGE,
     MINIMUM_IP_ECHO_SERVER_THREADS,
 };
-use ip_echo_server::{IpEchoServerMessage, IpEchoServerResponse};
+use {
+    ip_echo_client::{
+        do_verify_reachable_ports, ip_echo_server_request, ip_echo_server_request_with_binding,
+    },
+    ip_echo_server::IpEchoServerMessage,
+};
 
 /// A data type representing a public Udp socket
 pub struct UdpSocketPair {
@@ -46,77 +42,10 @@ pub const MINIMUM_VALIDATOR_PORT_RANGE_WIDTH: u16 = 17; // VALIDATOR_PORT_RANGE 
 pub(crate) const HEADER_LENGTH: usize = 4;
 pub(crate) const IP_ECHO_SERVER_RESPONSE_LENGTH: usize = HEADER_LENGTH + 23;
 
-async fn ip_echo_server_request(
-    ip_echo_server_addr: SocketAddr,
-    msg: IpEchoServerMessage,
-    bind_address: Option<IpAddr>,
-) -> anyhow::Result<IpEchoServerResponse> {
-    let timeout = Duration::new(5, 0);
-    let socket = tokio::net::TcpSocket::new_v4()?;
-    if let Some(addr) = bind_address {
-        socket.bind(SocketAddr::new(addr, 0))?;
-    }
-
-    async fn do_make_request(
-        socket: TcpSocket,
-        ip_echo_server_addr: SocketAddr,
-        msg: IpEchoServerMessage,
-    ) -> anyhow::Result<BytesMut> {
-        let mut stream = socket.connect(ip_echo_server_addr).await?;
-        // Start with HEADER_LENGTH null bytes to avoid looking like an HTTP GET/POST request
-        let mut bytes = BytesMut::with_capacity(IP_ECHO_SERVER_RESPONSE_LENGTH);
-        bytes.extend_from_slice(&[0u8; HEADER_LENGTH]);
-        bytes.extend_from_slice(&bincode::serialize(&msg)?);
-
-        // End with '\n' to make this request look HTTP-ish and tickle an error response back
-        // from an HTTP server
-        bytes.put_u8(b'\n');
-        stream.write_all(&bytes).await?;
-        stream.flush().await?;
-
-        bytes.clear();
-        let _n = stream.read_buf(&mut bytes).await?;
-        stream.shutdown().await?;
-
-        Ok(bytes)
-    }
-
-    let response =
-        tokio::time::timeout(timeout, do_make_request(socket, ip_echo_server_addr, msg)).await??;
-    // It's common for users to accidentally confuse the validator's gossip port and JSON
-    // RPC port.  Attempt to detect when this occurs by looking for the standard HTTP
-    // response header and provide the user with a helpful error message
-    if response.len() < HEADER_LENGTH {
-        bail!("Response too short, received {} bytes", response.len());
-    }
-
-    let (response_header, body) =
-        response
-            .split_first_chunk::<HEADER_LENGTH>()
-            .ok_or(anyhow::anyhow!(
-                "Not enough data in the response from {ip_echo_server_addr}!"
-            ))?;
-    let payload = match response_header {
-        [0, 0, 0, 0] => bincode::deserialize(&response[HEADER_LENGTH..])?,
-        [b'H', b'T', b'T', b'P'] => {
-            let http_response = std::str::from_utf8(body);
-            match http_response {
-                Ok(r) => bail!("Invalid gossip entrypoint. {ip_echo_server_addr} looks to be an HTTP port replying with {r}"),
-                Err(_) => bail!("Invalid gossip entrypoint. {ip_echo_server_addr} looks to be an HTTP port."),
-            }
-        }
-        _ => {
-            bail!("Invalid gossip entrypoint. {ip_echo_server_addr} provided unexpected header bytes {response_header:?} ");
-        }
-    };
-
-    Ok(payload)
-}
-
 /// Determine the public IP address of this machine by asking an ip_echo_server at the given
 /// address
 pub fn get_public_ip_addr(ip_echo_server_addr: &SocketAddr) -> Result<IpAddr, String> {
-    let fut = ip_echo_server_request(*ip_echo_server_addr, IpEchoServerMessage::default(), None);
+    let fut = ip_echo_server_request(*ip_echo_server_addr, IpEchoServerMessage::default());
     let rt = tokio::runtime::Builder::new_current_thread()
         .enable_all()
         .build()
@@ -131,10 +60,10 @@ pub fn get_public_ip_addr_with_binding(
     ip_echo_server_addr: &SocketAddr,
     bind_address: IpAddr,
 ) -> anyhow::Result<IpAddr> {
-    let fut = ip_echo_server_request(
+    let fut = ip_echo_server_request_with_binding(
         *ip_echo_server_addr,
         IpEchoServerMessage::default(),
-        Some(bind_address),
+        bind_address,
     );
     let rt = tokio::runtime::Builder::new_current_thread()
         .enable_all()
@@ -143,8 +72,9 @@ pub fn get_public_ip_addr_with_binding(
     Ok(resp.address)
 }
 
+///Retrieves cluster shred version from Entrypoint address provided
 pub fn get_cluster_shred_version(ip_echo_server_addr: &SocketAddr) -> Result<u16, String> {
-    let fut = ip_echo_server_request(*ip_echo_server_addr, IpEchoServerMessage::default(), None);
+    let fut = ip_echo_server_request(*ip_echo_server_addr, IpEchoServerMessage::default());
     let rt = tokio::runtime::Builder::new_current_thread()
         .enable_all()
         .build()
@@ -154,218 +84,27 @@ pub fn get_cluster_shred_version(ip_echo_server_addr: &SocketAddr) -> Result<u16
         .ok_or_else(|| "IP echo server does not return a shred-version".to_owned())
 }
 
+///Retrieves cluster shred version from Entrypoint address provided, binds client-side socket to IP provided
 pub fn get_cluster_shred_version_with_binding(
     ip_echo_server_addr: &SocketAddr,
     bind_address: IpAddr,
 ) -> anyhow::Result<u16> {
-    let fut = ip_echo_server_request(
+    let fut = ip_echo_server_request_with_binding(
         *ip_echo_server_addr,
         IpEchoServerMessage::default(),
-        Some(bind_address),
+        bind_address,
     );
     let rt = tokio::runtime::Builder::new_current_thread()
         .enable_all()
         .build()?;
     let resp = rt.block_on(fut)?;
     resp.shred_version
-        .ok_or_else(|| anyhow!("IP echo server does not return a shred-version"))
-}
-// Checks if any of the provided TCP/UDP ports are not reachable by the machine at
-// `ip_echo_server_addr`
-const DEFAULT_TIMEOUT: Duration = Duration::from_secs(5);
-const DEFAULT_RETRY_COUNT: usize = 5;
-
-async fn do_verify_reachable_ports(
-    ip_echo_server_addr: SocketAddr,
-    tcp_listeners: Vec<(u16, TcpListener)>,
-    udp_sockets: &[&UdpSocket],
-    timeout: Duration,
-    udp_retry_count: usize,
-) -> bool {
-    info!(
-        "Checking that tcp ports {:?} are reachable from {:?}",
-        tcp_listeners, ip_echo_server_addr
-    );
-
-    let tcp_ports: Vec<_> = tcp_listeners.iter().map(|(port, _)| *port).collect();
-    let _ = ip_echo_server_request(
-        ip_echo_server_addr,
-        IpEchoServerMessage::new(&tcp_ports, &[]),
-        None,
-    )
-    .await
-    .map_err(|err| warn!("ip_echo_server request failed: {}", err));
-
-    let mut ok = true;
-    let mut checkers = Vec::new();
-
-    // since we do not know if tcp_listeners are nonblocking, we have to run them in native threads.
-    for (port, tcp_listener) in tcp_listeners {
-        let listening_addr = tcp_listener.local_addr().unwrap();
-        let (sender, receiver) = oneshot::channel();
-        let thread_handle = std::thread::Builder::new()
-            .name(format!("solVrfyTcp{port:05}"))
-            .spawn(move || {
-                debug!("Waiting for incoming connection on tcp/{}", port);
-                match tcp_listener.incoming().next() {
-                    Some(_) => {
-                        // ignore errors here since this can only happen if a timeout was detected.
-                        // timeout drops the receiver part of the channel resulting in failure to send.
-                        let _ = sender.send(());
-                    }
-                    None => warn!("tcp incoming failed"),
-                }
-            })
-            .unwrap();
-
-        // Set the timeout on the receiver
-        let receiver = tokio::time::timeout(timeout, receiver);
-        checkers.push((listening_addr, thread_handle, receiver));
-    }
-
-    for (listening_addr, thread_handle, receiver) in checkers {
-        match receiver.await {
-            Ok(Ok(_)) => {
-                info!("tcp/{} is reachable", listening_addr.port());
-            }
-            Ok(Err(_v)) => {
-                unreachable!("The receive on oneshot channel should never fail");
-            }
-            Err(_t) => {
-                error!(
-                    "Received no response at tcp/{}, check your port configuration",
-                    listening_addr.port()
-                );
-                // Ugh, std rustc doesn't provide accepting with timeout or restoring original
-                // nonblocking-status of sockets because of lack of getter, only the setter...
-                // So, to close the thread cleanly, just connect from here.
-                // ref: https://github.com/rust-lang/rust/issues/31615
-                TcpStream::connect_timeout(&listening_addr, timeout).unwrap();
-                ok = false;
-            }
-        }
-        thread_handle.join().expect("Thread should exit cleanly")
-    }
-
-    if !ok {
-        // No retries for TCP, abort on any failure
-        return false;
-    }
-
-    // now check UDP ports
-    let mut ok = true;
-    let mut udp_ports: BTreeMap<_, _> = BTreeMap::new();
-    udp_sockets.iter().for_each(|udp_socket| {
-        let port = udp_socket.local_addr().unwrap().port();
-        udp_ports
-            .entry(port)
-            .or_insert_with(Vec::new)
-            .push(udp_socket);
-    });
-    let udp_ports: Vec<_> = udp_ports.into_iter().collect();
-
-    info!(
-        "Checking that udp ports {:?} are reachable from {:?}",
-        udp_ports.iter().map(|(port, _)| port).collect::<Vec<_>>(),
-        ip_echo_server_addr
-    );
-
-    'outer: for checked_ports_and_sockets in udp_ports.chunks(MAX_PORT_COUNT_PER_MESSAGE) {
-        ok = false;
-
-        for udp_remaining_retry in (0_usize..udp_retry_count).rev() {
-            let (checked_ports, checked_socket_iter) = (
-                checked_ports_and_sockets
-                    .iter()
-                    .map(|(port, _)| *port)
-                    .collect::<Vec<_>>(),
-                checked_ports_and_sockets
-                    .iter()
-                    .flat_map(|(_, sockets)| sockets),
-            );
-
-            let _ = ip_echo_server_request(
-                ip_echo_server_addr,
-                IpEchoServerMessage::new(&[], &checked_ports),
-                None,
-            )
-            .await
-            .map_err(|err| warn!("ip_echo_server request failed: {}", err));
-
-            // Spawn threads at once!
-            let reachable_ports = Arc::new(RwLock::new(HashSet::new()));
-            let thread_handles: Vec<_> = checked_socket_iter
-                .map(|udp_socket| {
-                    let port = udp_socket.local_addr().unwrap().port();
-                    let udp_socket = udp_socket.try_clone().expect("Unable to clone udp socket");
-                    let reachable_ports = reachable_ports.clone();
-
-                    std::thread::Builder::new()
-                        .name(format!("solVrfyUdp{port:05}"))
-                        .spawn(move || {
-                            let start = Instant::now();
-
-                            let original_read_timeout = udp_socket.read_timeout().unwrap();
-                            udp_socket
-                                .set_read_timeout(Some(Duration::from_millis(250)))
-                                .unwrap();
-                            loop {
-                                if reachable_ports.read().unwrap().contains(&port)
-                                    || Instant::now().duration_since(start) >= timeout
-                                {
-                                    break;
-                                }
-
-                                let recv_result = udp_socket.recv(&mut [0; 1]);
-                                debug!(
-                                    "Waited for incoming datagram on udp/{}: {:?}",
-                                    port, recv_result
-                                );
-
-                                if recv_result.is_ok() {
-                                    reachable_ports.write().unwrap().insert(port);
-                                    break;
-                                }
-                            }
-                            udp_socket.set_read_timeout(original_read_timeout).unwrap();
-                        })
-                        .unwrap()
-                })
-                .collect();
-
-            // Now join threads!
-            // Separate from the above by collect()-ing as an intermediately step to make the iterator
-            // eager not lazy so that joining happens here at once after creating bunch of threads
-            // at once.
-            for thread in thread_handles {
-                thread.join().unwrap();
-            }
-
-            let reachable_ports = reachable_ports.read().unwrap().clone();
-            if reachable_ports.len() == checked_ports.len() {
-                info!(
-                    "checked udp ports: {:?}, reachable udp ports: {:?}",
-                    checked_ports, reachable_ports
-                );
-                ok = true;
-                break;
-            } else if udp_remaining_retry > 0 {
-                // Might have lost a UDP packet, retry a couple times
-                error!(
-                    "checked udp ports: {:?}, reachable udp ports: {:?}",
-                    checked_ports, reachable_ports
-                );
-                error!("There are some udp ports with no response!! Retrying...");
-            } else {
-                error!("Maximum retry count is reached....");
-                break 'outer;
-            }
-        }
-    }
-
-    ok
+        .ok_or_else(|| anyhow::anyhow!("IP echo server does not return a shred-version"))
 }
 
+/// Checks if all of the provided TCP/UDP ports are not reachable by the machine at
+/// `ip_echo_server_addr`. Tests must complete within timeout provided.
+/// Tests will run in parallel when possible.
 pub fn verify_reachable_ports(
     ip_echo_server_addr: &SocketAddr,
     tcp_listeners: Vec<(u16, TcpListener)>,
@@ -379,8 +118,8 @@ pub fn verify_reachable_ports(
         *ip_echo_server_addr,
         tcp_listeners,
         udp_sockets,
-        DEFAULT_TIMEOUT,
-        DEFAULT_RETRY_COUNT,
+        ip_echo_client::TIMEOUT,
+        ip_echo_client::DEFAULT_RETRY_COUNT,
     );
     rt.block_on(fut)
 }
@@ -860,7 +599,12 @@ pub fn bind_more_with_config(
 
 #[cfg(test)]
 mod tests {
-    use {super::*, std::net::Ipv4Addr, tokio::runtime::Runtime};
+    use {
+        super::*,
+        ip_echo_server::IpEchoServerResponse,
+        std::{net::Ipv4Addr, time::Duration},
+        tokio::runtime::Runtime,
+    };
 
     fn runtime() -> Runtime {
         tokio::runtime::Builder::new_current_thread()
